+++ domain/src/main/java/com/ntros/mprocswift/model/transactions/card/AuthorizedHold.java
@@ -5,7 +5,6 @@ import jakarta.persistence.*;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Entity
@@ -13,21 +12,21 @@ import java.time.OffsetDateTime;
 @RequiredArgsConstructor
 public class AuthorizedHold {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer authorizedHoldId;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Integer authorizedHoldId;
 
-    @ManyToOne
-    @JoinColumn(name = "card_authorization_id")
-    private CardAuthorization cardAuthorization;
+  @ManyToOne
+  @JoinColumn(name = "card_authorization_id")
+  private CardAuthorization cardAuthorization;
 
-    @ManyToOne
-    @JoinColumn(name = "wallet_id")
-    private Wallet wallet;
+  @ManyToOne
+  @JoinColumn(name = "wallet_id")
+  private Wallet wallet;
 
-    private BigDecimal holdAmount;
-    private OffsetDateTime holdDate;
-    private OffsetDateTime expiresAt;
-    private Boolean isReleased;
-    private OffsetDateTime releasedAt;
+  private long holdAmount;
+  private OffsetDateTime holdDate;
+  private OffsetDateTime expiresAt;
+  private Boolean isReleased;
+  private OffsetDateTime releasedAt;
 }
