+++ service/src/main/java/com/ntros/mprocswift/service/ledger/Posting.java
@@ -1,13 +1,10 @@
 package com.ntros.mprocswift.service.ledger;
 
-import com.ntros.mprocswift.model.currency.Currency;
 import com.ntros.mprocswift.model.ledger.LedgerAccount;
-
-import java.math.BigDecimal;
 
 public record Posting(
     LedgerAccount debitAccount,
     LedgerAccount creditAccount,
-    BigDecimal amount,
+    long amount,
     String description,
     String entryGroupKey) {}
